Reconcile #4983 cache.ts with upstream: split runner-count caches into runner-count-cache.ts
- Restore upstream githubCache in cache.ts (was clobbered by #4983)
- Move ec2RunnerCountCache + dynamoDbRunnerCountCache to runner-count-cache.ts
- Drop duplicated githubCache/GhRunners/Octokit import
- 21/21 count-cache tests pass
